feat(statistics): 5천만원 정도 구입하니 오버플로우 발생 취약. 총 상금 long 타입으로 변경

+++ src/main/java/lotto/domain/Statistics.java
@@ -37,10 +37,10 @@ public class Statistics {
         }
     }
 
-    public int calcTotalPrize() {
+    public long calcTotalPrize() {
         return buildWinningCounts().entrySet()
                 .stream()
-                .mapToInt(
+                .mapToLong(
                         entry -> entry.getKey()
                                 .calcPrize(
                                         entry.getValue()
@@ -50,7 +50,7 @@ public class Statistics {
     }
 
     public double calcYield(int amount) {
-        int totalPrize = calcTotalPrize();
+        long totalPrize = calcTotalPrize();
         return 100 * totalPrize / (double) amount;
     }
